The calculation for the delay was also wrong! Fck rounding

+++ Morser/Model/VibrationEngine.swift
@@ -108,7 +108,7 @@ class VibrationEngine: ObservableObject {
             for _ in 1...3 {
 //                playHaptics(url: dashAhapUrl!)
                 playHapticsFile(named: "dash")
-                usleep(UInt32(dashDuration) * (10000 * UInt32(VibrationEngine.timeUnit)))
+                usleep(UInt32(dashDuration * 10000.0 * VibrationEngine.timeUnit))
             }
             vibrationTimer = Timer.scheduledTimer(withTimeInterval: dashDuration + (character == nextCharacter ? sameCharacterSeparatorDelay : characterSeparatorDelay), repeats: false) { _ in
                 self.triggerNextVibration()
